+++ app/controllers/bazaar_admin/warehouse_sku_country_admin_controller.rb
@@ -1,0 +1,5 @@
+module BazaarAdmin
+	class WarehouseSkuCountryAdminController < BazaarAdmin::EcomAdminController
+
+	end
+end