+++ app/helpers/bazaar/application_helper.rb
@@ -1,7 +1,0 @@
-module Bazaar
-  module ApplicationHelper
-		def humanize_first
-			return lambda { |e| e.first.humanize.titleize }
-		end
-  end
-end