+++ app/helpers/bazaar_admin/application_helper.rb
@@ -1,0 +1,7 @@
+module BazaarAdmin
+  module ApplicationHelper
+		def humanize_first
+			return lambda { |e| e.first.humanize.titleize }
+		end
+  end
+end