+++ lib/bazaar_admin.rb
@@ -1,0 +1,4 @@
+require "bazaar_admin/engine"
+
+module BazaarAdmin
+end