+++ lib/tasks/install_files/elastic_searchable_concerns/bazaar/discount_searchable.rb
@@ -1,4 +1,4 @@
-module Bazaar
+module BazaarAdmin
 	module DiscountSearchable
 		extend ActiveSupport::Concern