+++ lib/tasks/install_files/elastic_searchable_concerns/bazaar/order_searchable.rb
@@ -1,4 +1,4 @@
-module Bazaar
+module BazaarAdmin
 	module OrderSearchable
 		extend ActiveSupport::Concern