+++ lib/tasks/install_files/elastic_searchable_concerns/bazaar/product_searchable.rb
@@ -1,4 +1,4 @@
-module Bazaar
+module BazaarAdmin
 	module ProductSearchable
 		extend ActiveSupport::Concern