+++ lib/tasks/install_files/elastic_searchable_concerns/bazaar/subscription_searchable.rb
@@ -1,4 +1,4 @@
-module Bazaar
+module BazaarAdmin
 	module SubscriptionSearchable
 		extend ActiveSupport::Concern