+++ lib/tasks/install_files/subscription_plan_searchable.rb
@@ -1,4 +1,4 @@
-module Bazaar
+module BazaarAdmin
 	module SubscriptionPlanSearchable
 		extend ActiveSupport::Concern